backend/pyproject.toml
[tool.poetry]
name = "FastAPI Server"
version = "0.1.0"
description = ""
authors = ["Nikolay <nchernov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.25.1"
uvicorn = "^0.14.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^0.18.0"
python-multipart = "^0.0.5"
bcrypt = "^3.2.0"
asyncpg = "^0.27.0"
SQLAlchemy = "^1.4.36"
email-validator = "^1.2.1"
psycopg2-binary = "^2.9.3"
alembic = "^1.7.7"
invoke = "^1.7.1"
ipython = "^8.3.0"
fastapi = "^0.95.1"
fastapi-pagination = "^0.12.2"
pydantic = "^1.10.7"
httpx = "^0.24.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pytest = "^6.2.4"
black = "^23.1.0"
asynctest = "^0.13.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
